code/server/wan_local_service/requirements-service.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
websocket-client==1.8.0
